+++ RazorSvelte/App/404.svelte
@@ -3,9 +3,8 @@
 </script>
 
 <Layout>
-    <div class="container text-center">
-        <h1 class="text-danger">404 Not Found</h1>
-        
+    <div class="bx--content bx--toast-notification--error">
+        <h1 class="">404 Not Found</h1>
         <p>
             This page doesn't seem to exists.
         </p>