+++ RazorSvelte/App/About.svelte
@@ -1,23 +1,16 @@
 <script lang="ts">
-    import Paper, { Title, Content } from '@smui/paper';
-    import Layout from "./Shared/Layout.svelte";
-    
+import { ClickableTile } from "carbon-components-svelte";
+import Layout from "./Shared/Layout.svelte";
 </script>
 
 <Layout>
-    <Paper color="primary" variant="outlined" style="margin-top: 25px">
-        <Title>
+    <ClickableTile href="https://github.com/vb-consulting/Norm.net">
+        <h1>
             <code>Norm.net</code>
-        </Title>
-        <Content>
-            <p>
-                High performance micro-ORM database mapper and modernized Dapper replacement for .NET Standard 2.1 and higher
-            </p>
-            <p>
-                Check it out on <a href="https://github.com/vb-consulting/Norm.net">GitHub</a>
-            </p>
-        </Content>
-    </Paper>
+        </h1>
+        <p>
+            High performance micro-ORM database mapper and modernized Dapper replacement for .NET Standard 2.1 and higher
+        </p>
+        
+    </ClickableTile>
 </Layout>
-
-