+++ RazorSvelte/Pages/Shared/_Loading.cshtml.cs
@@ -1,9 +1,0 @@
-﻿namespace RazorSvelte.Pages.Shared
-{
-    public class Loading
-    {
-        public string Id { get; init; } = "loading";
-        public string MarginTop { get; init; } = "100px";
-        public string Size { get; init; } = "10rem";
-    }
-}