+++ RazorSvelte/Scripts/build-all.js
@@ -45,9 +45,7 @@ const removeMaps = () => new Promise(resolve => {
 
 const promises = [];
 
-if (!process.argv || process.argv.length == 0 || process.argv.indexOf("--skip-theme-compile") == -1) {
-    promises.push(exec(`npm run fe-theme-compile-all`));
-}
+promises.push(exec(`npm run fe-scss-build`));
 
 for (let config of getAllConfigs("./Pages")) {
     promises.push(exec("rollup -c " + config));