+++ RazorSvelte/Scripts/rollup.config.js
@@ -6,6 +6,7 @@ import sveltePreprocess from "svelte-preprocess";
 import typescript from "@rollup/plugin-typescript";
 import css from "rollup-plugin-css-only";
 import replace from '@rollup/plugin-replace';
+import { optimizeImports } from "carbon-preprocess-svelte";
 
 const production = !process.env.ROLLUP_WATCH;
 
@@ -47,7 +48,7 @@ export default (param, globals) => {
                 "process.env.NODE_ENV": JSON.stringify("development")
             }),
             svelte({
-                preprocess: sveltePreprocess({ sourceMap: !production }),
+                preprocess: [sveltePreprocess({ sourceMap: !production }), optimizeImports()],
                 compilerOptions: {
                     // enable run-time checks when not in production
                     dev: !production
@@ -61,8 +62,6 @@ export default (param, globals) => {
                     handler(warning);
                 },
             }),
-            // we"ll extract any component CSS out into
-            // a separate file - better for performance
             css({ output: cssOutput }),
     
             // If you have external dependencies installed from